Refactor Wujie child components and update link handling
