+++ src/pages/demo/wujie/a-projects/wujie-child-2/src/SpecialLink.tsx
@@ -2,25 +2,27 @@ import { Fragment, useEffect, useState } from "react";
 import { Link } from "./App";
 
 export const SpecialLink = () => {
-  const links = window.$wujie.props?.api.getLinks();
-  // console.log("links", links);
+  const [allLinks, setAllLinks] = useState<
+    {
+      url: string;
+      name: string;
+    }[]
+  >([]);
 
-  const [linkList, setLinkList] = useState(
-    links?.filter((item) => item.name.includes("github"))
-  );
-
-  // filter仅数组中当链接包括 github 时，显示该链接
-  // const githubLink = ;
+  const githubLink = allLinks?.filter((item) => item.name.includes("github"));
 
   const [count, setCount] = useState(0);
 
   const handleUpdate = (data: any) => {
     console.log("handleUpdate", data);
-    setLinkList(data.filter((item: any) => item.name.includes("github")));
-    // setCount((count) => count + 1);
+    setAllLinks(data);
   };
 
   useEffect(() => {
+    const links = window.$wujie.props?.api.getLinks();
+
+    setAllLinks(links!);
+
     window.$wujie?.bus.$on("contextUpdated", handleUpdate);
     return () => {
       window.$wujie?.bus.$off("contextUpdated", handleUpdate);
@@ -33,11 +35,17 @@ export const SpecialLink = () => {
         display: "flex",
         flexWrap: "wrap",
         alignItems: "center",
-        justifyContent: "center",
       }}
     >
-      {linkList &&
-        linkList.map((item, index) => (
+      <div
+        style={{
+          color: "red",
+        }}
+      >
+        I'm alive in iframe
+      </div>
+      {githubLink &&
+        githubLink.map((item, index) => (
           <Fragment key={index}>
             <Link link={item.url} name={item.name + "+count:" + count} />
             <button onClick={() => setCount((count) => count + 1)}>
