+++ src/pages/demo/wujie/a-projects/wujie-child-2/tsconfig.node.tsbuildinfo
