+++ src/pages/demo/wujie/a-projects/wujie-child/package.json
